+++ supervisor-parent/supervisor-dao/src/main/java/com/sxj/supervisor/entity/contract/ModifyItemEntity.java
@@ -55,7 +55,7 @@ public class ModifyItemEntity extends Pagable implements Serializable
      * 数量
      **/
     @Column(name = "QUANTITY")
-    private Float quantity;
+    private Double quantity;
     
     /**
      * 单价
@@ -131,12 +131,12 @@ public class ModifyItemEntity extends Pagable implements Serializable
         this.productModel = productModel;
     }
     
-    public Float getQuantity()
+    public Double getQuantity()
     {
         return quantity;
     }
     
-    public void setQuantity(Float quantity)
+    public void setQuantity(Double quantity)
     {
         this.quantity = quantity;
     }
